updating tests for better organization and added more
